JsApiService: Refactor, make implementation internal.
